Add CORS and Authorize token

+++ ReviewBook.API/Controllers/AuthorController.cs
@@ -1,4 +1,8 @@
+using System.Net;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Net.Http.Headers;
 using ReviewBook.API.Data.Entities;
 using ReviewBook.API.DTOs;
 using ReviewBook.API.Services;
@@ -10,40 +14,63 @@ namespace ReviewBook.API.Controllers
     public class AuthorController : ControllerBase
     {
         private readonly IAuthorService _AuthorService;
+        private readonly IUserService _userService;
 
-        public AuthorController(IAuthorService authorService)
+        public AuthorController(IAuthorService authorService, IUserService userService)
         {
             _AuthorService = authorService;
+            _userService = userService;
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpGet]
         public ActionResult<IEnumerable<Author>> Get()
         {
-            return Ok(_AuthorService.GetAllAuthors());
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_AuthorService.GetAllAuthors());
+            return BadRequest("Không đủ quyền");
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpGet("{id}")]
         public ActionResult<Author> Get(int id)
         {
             return Ok(_AuthorService.GetAuthorById(id));
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPost]
         public ActionResult<Author> Post([FromBody] CreateAuthorDTOs value)
         {
-            return Ok(_AuthorService.CreateAuthor(value.toAuthorEntity()));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_AuthorService.CreateAuthor(value.toAuthorEntity()));
+            return BadRequest("Không đủ quyền");
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPut("{id}")]
         public ActionResult<Author> Put(int id, [FromBody] UpdateInforAuthorDTOs value)
         {
-            return Ok(_AuthorService.UpdateAuthor(value.toAuthorEntity(id)));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_AuthorService.UpdateAuthor(value.toAuthorEntity(id)));
+            return BadRequest("Không đủ quyền");
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
         {
-            return Ok(_AuthorService.DeleteAuthor(id));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_AuthorService.DeleteAuthor(id));
+            return BadRequest("Không đủ quyền");
         }
     }
 }
