+++ ReviewBook.API/Controllers/PublisherController.cs
@@ -1,25 +1,35 @@
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Net.Http.Headers;
 using ReviewBook.API.Data.Entities;
 using ReviewBook.API.DTOs;
 using ReviewBook.API.Services;
 
 namespace ReviewBook.API.Controllers
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     [Route("api/[controller]")]
     [ApiController]
     public class PublisherController : ControllerBase
     {
         private readonly IPublisherService _PublisherService;
+        private readonly IUserService _userService;
 
-        public PublisherController(IPublisherService publisherService)
+        public PublisherController(IPublisherService publisherService, IUserService userService)
         {
             _PublisherService = publisherService;
+            _userService = userService;
         }
 
         [HttpGet]
         public ActionResult<IEnumerable<Publisher>> Get()
         {
-            return Ok(_PublisherService.GetAllPublishers());
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_PublisherService.GetAllPublishers());
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpGet("{id}")]
@@ -31,19 +41,31 @@ namespace ReviewBook.API.Controllers
         [HttpPost]
         public ActionResult<Publisher> Post([FromBody] CreatePublisherDTOs value)
         {
-            return Ok(_PublisherService.CreatePublisher(value.toAuthorEntity()));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_PublisherService.CreatePublisher(value.toAuthorEntity()));
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpPut("{id}")]
         public ActionResult<Publisher> Put(int id, [FromBody] UpdateInforPublisherDTOs value)
         {
-            return Ok(_PublisherService.UpdatePublisher(value.toPublisherEntity(id)));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_PublisherService.UpdatePublisher(value.toPublisherEntity(id)));
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
         {
-            return Ok(_PublisherService.DeletePublisher(id));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_PublisherService.DeletePublisher(id));
+            return BadRequest("Không đủ quyền");
         }
     }
 }
