+++ ReviewBook.API/Controllers/TagController.cs
@@ -1,49 +1,75 @@
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Net.Http.Headers;
 using ReviewBook.API.Data.Entities;
 using ReviewBook.API.DTOs;
 using ReviewBook.API.Services;
 
 namespace ReviewBook.API.Controllers
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     [Route("api/[controller]")]
     [ApiController]
     public class TagController : ControllerBase
     {
         private readonly ITagService _TagService;
+        private readonly IUserService _userService;
 
-        public TagController(ITagService tagService)
+        public TagController(ITagService tagService, IUserService userService)
         {
             _TagService = tagService;
+            _userService = userService;
         }
 
         [HttpGet]
         public ActionResult<IEnumerable<Tag>> Get()
         {
-            return Ok(_TagService.GetAllTags());
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_TagService.GetAllTags());
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpGet("{id}")]
         public ActionResult<Tag> Get(int id)
         {
-            return Ok(_TagService.GetTagById(id));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_TagService.GetTagById(id));
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpPost]
         public ActionResult<Tag> Post([FromBody] CreateTagDTOs value)
         {
-            return Ok(_TagService.CreateTag(value.toTagEntity()));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_TagService.CreateTag(value.toTagEntity()));
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpPut("{id}")]
         public ActionResult<Tag> Put(int id, [FromBody] UpdateTagDTOs value)
         {
-            return Ok(_TagService.UpdateTag(value.toTagEntity(id)));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_TagService.UpdateTag(value.toTagEntity(id)));
+            return BadRequest("Không đủ quyền");
         }
 
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
         {
-            return Ok(_TagService.DeleteTag(id));
+            var _bearer_token = Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
+            var acc = _userService.jwtTokenToAccount(_bearer_token);
+            if (acc.ID_Role == 1)
+                return Ok(_TagService.DeleteTag(id));
+            return BadRequest("Không đủ quyền");
         }
     }
 }
